+++ src/kete/__init__.py
@@ -6,12 +6,11 @@ from . import (
     cache,
     constants,
     covariance,
-    fitting,
     flux,
     fov,
-    gaia,
     mpc,
     neos,
+    orbit_fitting,
     plot,
     population,
     ptf,
@@ -91,7 +90,6 @@ __all__ = [
     "constants",
     "covariance",
     "Data",
-    "fitting",
     "flux",
     "flux_to_mag",
     "fov",
@@ -106,6 +104,7 @@ __all__ = [
     "mpc",
     "neos",
     "NeosCmos",
+    "orbit_fitting",
     "NeosVisit",
     "OmniDirectionalFOV",
     "plot",
